Add SHEETPATH text variable
This is referenced by the custom title sheet and set for schematics.
Inside KiCad, the pcb editor just leaves this empty, but it seems kibot
just leaves the variable name when generating a PCB PDF.

This sets an explicit text variable into the project as a fallback for
this case.