For DRC, ignore unconnected nets and warnings
In general, we should not need to fail on DRC warnings, so ignore them
(specific board can revert this again if they are DRC-warning free).

Note that this *also* ignores unconnected nets, which is unwanted, but
currently kibot does not allow controlling these separately. See
https://github.com/INTI-CMNB/KiBot/issues/374

Because this reduces the need for skipping DRC entirely, also remove the
commented example for that from the kibot-project-config.yaml example.