Grant permissions to create releases
Originally, the default GITHUB_TOKEN seems to have defaulted to write
permissions, but now it seems the default (in organization and/or
repository settings) is read-only access.

By declaring the needed permission in the workflow job, additional
permission is given. This must be done in the main workflow, it cannot
be done in a called workflow, so this is why this is done in the
example, not in the actual (called) workflow.