Include project files in outputs
This generates KiCad project files with the revision variables set
to the auto-detected values (e.g. from the git tag or branch when
running in a workflow). This produces a KiCad project that can be used
for assembly. It also includes all used symbols, footprints and 3D
models to make a fully self-contained project for archival purposes.